+++ Demo/Demo/SwiftUIComponents/PayPalWebPayments/PayPalWebButtonsView.swift
@@ -20,14 +20,14 @@ struct PayPalWebButtonsView: View {
                 .font(.headline)
                 Picker("Funding Source", selection: $selectedFundingSource) {
                     Text("PayPal").tag(PayPalWebCheckoutFundingSource.paypal)
-                    Text("PayPal Credit").tag(PayPalWebCheckoutFundingSource.paylater)
-                    Text("Pay Later").tag(PayPalWebCheckoutFundingSource.paypalCredit)
+                    Text("PayPal Credit").tag(PayPalWebCheckoutFundingSource.paypalCredit)
+                    Text("Pay Later").tag(PayPalWebCheckoutFundingSource.paylater)
                 }
                 .pickerStyle(SegmentedPickerStyle())
 
                 switch selectedFundingSource {
                 case .paypalCredit:
-                    PayPalCreditButton.Representable(color: .black, size: .full) {
+                    PayPalCreditButton.Representable(color: .gold, size: .full) {
                         payPalWebViewModel.paymentButtonTapped(funding: .paypalCredit)
                     }
                 case .paylater: