+++ Sources/PaymentButtons/PayPalCreditButton.swift
@@ -8,12 +8,13 @@ public final class PayPalCreditButton: PaymentButton {
     Available colors for PayPalCreditButton.
     */
     public enum Color: String {
+        case gold
         case white
         case black
         case darkBlue
 
         var color: PaymentButtonColor {
-            PaymentButtonColor(rawValue: rawValue) ?? .darkBlue
+            PaymentButtonColor(rawValue: rawValue) ?? .white
         }
     }